# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Veel V2 API
  version: 0.1.0
  description: >
    Contract-first API source of truth for the Veel v2 standalone build. Fastify routes,
    generated clients, tests, and docs must stay aligned with this file.
servers:
  - url: https://api.veel.example
    description: Production
  - url: http://localhost:4000
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /v1/events/{eventId}/access-passes/intents:
    post:
      operationId: createAccessPassIntent
      tags: [Events]
      summary: Create Event Access Pass payment or approval intent
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/RequiredIdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/CreateAccessPassIntent"
      responses:
        "201": { $ref: "#/components/responses/AccessPassIntent" }
  /v1/events/{eventId}/access-passes/requests:
    post:
      operationId: requestPrivateEventAccessPass
      tags: [Events]
      summary: Request creator approval for a private Event Access Pass
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/RequiredIdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/CreateAccessPassRequest"
      responses:
        "201": { $ref: "#/components/responses/AccessPassRequest" }
  /v1/access-passes/{accessPassId}/check-in:
    post:
      operationId: checkInAccessPass
      tags: [Access Passes]
      summary: Check in an Event Access Pass by backend-issued QR/token
      parameters:
        - $ref: "#/components/parameters/AccessPassId"
        - $ref: "#/components/parameters/RequiredIdempotencyKey"
      requestBody:
        $ref: "#/components/requestBodies/CheckInAccessPass"
      responses:
        "200": { $ref: "#/components/responses/AccessPass" }
...
  /v1/admin/events:
    get:
      operationId: listAdminEvents
      tags: [Admin]
      summary: Admin event queue and health
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": { $ref: "#/components/responses/EventPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/event-access-passes:
    get:
      operationId: listAdminEventAccessPasses
      tags: [Admin]
      summary: Admin Event Access Pass entitlement and check-in list
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": { $ref: "#/components/responses/AccessPassPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /v1/admin/media/assets:
    get:
      operationId: listAdminMediaAssets
      tags: [Admin]
      summary: Admin Bunny media asset provider projection
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": { $ref: "#/components/responses/AdminMediaAssetPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/age-kyc/age-checks:
    get:
      operationId: listAdminAgeChecks
      tags: [Admin]
      summary: Admin age assurance provider projection
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": { $ref: "#/components/responses/AdminAgeCheckPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/age-kyc/identity-checks:
    get:
      operationId: listAdminIdentityChecks
      tags: [Admin]
      summary: Admin identity, KYC, and KYB verification projection
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": { $ref: "#/components/responses/AdminIdentityCheckPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/ai/sessions:
    get:
      operationId: listAdminAiSessions
      tags: [Admin]
      summary: Admin AI scoped assistant sessions
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": { $ref: "#/components/responses/AdminAiSessionPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/ai/tool-calls:
    get:
      operationId: listAdminAiToolCalls
      tags: [Admin]
      summary: Admin AI tool call audit projection
      parameters:
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200": { $ref: "#/components/responses/AdminAiToolCallPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/mutuals/safety:
    get:
      operationId: getAdminMutualsSafety
      tags: [Admin]
      summary: Mutuals safety queue and metrics
      responses:
        "200": { $ref: "#/components/responses/AdminMutualsSafety" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequiredIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 12, maxLength: 128 }
      description: Required for money, entitlement, Event Access, message, Mutuals, age, wallet, moderation, and admin mutations.
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
    EventId: { name: eventId, in: path, required: true, schema: { type: string, format: uuid } }
    AccessPassId: { name: accessPassId, in: path, required: true, schema: { type: string, format: uuid } }
...
  requestBodies:
...
    CreateAccessPassIntent: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateAccessPassIntentRequest" } } } }
    CreateAccessPassRequest: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateAccessPassRequestRequest" } } } }
    CheckInAccessPass: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CheckInAccessPassRequest" } } } }
...
  responses:
...
    Forbidden: { description: Forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
    EventPage: { description: Events, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/Event" } }, nextCursor: { type: string, nullable: true } } } } } }
    AccessPassIntent: { description: Event Access Pass intent, content: { application/json: { schema: { $ref: "#/components/schemas/AccessPassIntent" } } } }
    AccessPassRequest: { description: Event Access Pass request, content: { application/json: { schema: { $ref: "#/components/schemas/AccessPassRequest" } } } }
    AccessPass: { description: Event Access Pass, content: { application/json: { schema: { $ref: "#/components/schemas/AccessPass" } } } }
    AccessPassPage: { description: Event Access Passes, content: { application/json: { schema: { $ref: "#/components/schemas/AccessPassPage" } } } }
...
    AdminMutualsSafety: { description: Admin Mutuals safety, content: { application/json: { schema: { $ref: "#/components/schemas/AdminMutualsSafety" } } } }
...
    AdminMediaAssetPage: { description: Admin media provider projections, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/AdminMediaAsset" } }, nextCursor: { type: string, nullable: true } } } } } }
    AdminAgeCheckPage: { description: Admin age assurance checks, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/AdminAgeCheck" } }, nextCursor: { type: string, nullable: true } } } } } }
    AdminIdentityCheckPage: { description: Admin identity, KYC, and KYB checks, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/AdminIdentityCheck" } }, nextCursor: { type: string, nullable: true } } } } } }
    AdminAiSessionPage: { description: Admin AI sessions, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/AdminAiSession" } }, nextCursor: { type: string, nullable: true } } } } } }
    AdminAiToolCallPage: { description: Admin AI tool calls, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: "#/components/schemas/AdminAiToolCall" } }, nextCursor: { type: string, nullable: true } } } } } }
...
  schemas:
...
    ErrorCode:
      type: string
      enum: [unauthorized, forbidden, not_found, validation_failed, rate_limited, conflict, idempotency_conflict, provider_unavailable, payment_not_confirmed, webhook_signature_invalid]
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    ProductType:
      type: string
      enum: [tip, support, content_unlock, paid_message, live_pass, event_access_pass, creator_subscription, platform_subscription]
    Currency:
      type: string
      enum: [SOL, USDC]
...
    AgeState: { type: string, enum: [not_required, required, pending, verified, failed] }
...
    EventAccessPassType:
      type: object
      required: [id, label, priceMinor, currency, capacity, remaining, state]
      properties:
        id: { type: string, format: uuid }
        label: { type: string, minLength: 1, maxLength: 80 }
        priceMinor: { type: integer, nullable: true }
        currency: { $ref: "#/components/schemas/Currency" }
        capacity: { type: integer, minimum: 1 }
        remaining: { type: integer, minimum: 0 }
        state: { type: string, enum: [active, paused, sold_out] }
        saleStartsAt: { type: string, format: date-time, nullable: true }
        saleEndsAt: { type: string, format: date-time, nullable: true }
        perUserLimit: { type: integer, minimum: 1, maximum: 20, nullable: true }
...
    EventLocationDraft:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [digital_live_stream, physical] }
        label: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        provider: { type: string, enum: [openstreetmap, manual, none] }
        providerPlaceId: { type: string }
...
    PaymentIntent:
      type: object
      required: [id, productType, amountMinor, currency, state]
      properties:
        id: { type: string, format: uuid }
        productType: { $ref: "#/components/schemas/ProductType" }
        amountMinor: { type: integer }
        currency: { $ref: "#/components/schemas/Currency" }
        state: { type: string, enum: [pending, transaction_requested, submitted, confirmed, failed, expired] }
...
    Event:
      type: object
      required: [id, title, startsAt, accessRule, location, state, accessPassTypes]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, nullable: true }
        accessRule: { type: string, enum: [public_sale, private_apply] }
        location: { $ref: "#/components/schemas/EventLocationDraft" }
        state: { type: string, enum: [draft, published, sold_out, cancelled, completed] }
        accessPassTypes:
          type: array
          items: { $ref: "#/components/schemas/EventAccessPassType" }
    CreateAccessPassIntentRequest:
      type: object
      required: [accessPassTypeId]
      properties:
        accessPassTypeId: { type: string, format: uuid }
    AccessPassIntent:
      type: object
      required: [state]
      properties:
        state: { type: string, enum: [free_granted, approval_required, payment_required] }
        paymentIntent: { $ref: "#/components/schemas/PaymentIntent" }
        accessPass: { $ref: "#/components/schemas/AccessPass" }
    CreateAccessPassRequestRequest:
      type: object
      required: [accessPassTypeId]
      properties:
        accessPassTypeId: { type: string, format: uuid }
        note: { type: string, maxLength: 500 }
    AccessPassRequest:
      type: object
      required: [id, eventId, accessPassTypeId, state, createdAt]
      properties:
        id: { type: string, format: uuid }
        eventId: { type: string, format: uuid }
        accessPassTypeId: { type: string, format: uuid }
        state: { type: string, enum: [requested, approved, rejected, expired] }
        createdAt: { type: string, format: date-time }
    CheckInAccessPassRequest:
      type: object
      required: [qrToken]
      properties:
        qrToken: { type: string }
    AccessPass:
      type: object
      required: [id, eventId, accessPassTypeId, holderUserId, state, qrToken, createdAt]
      properties:
        id: { type: string, format: uuid }
        eventId: { type: string, format: uuid }
        accessPassTypeId: { type: string, format: uuid }
        holderUserId: { type: string, format: uuid }
        paymentIntentId: { type: string, format: uuid, nullable: true }
        state: { type: string, enum: [active, checked_in, revoked, expired] }
        qrToken: { type: string }
        checkedInAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    AccessPassPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AccessPass" } }
        nextCursor: { type: string, nullable: true }
...
    AiToolName:
      type: string
      enum:
        - explain_app_state
        - summarize_own_activity
        - find_own_purchases
        - draft_caption
        - suggest_hashtags
        - prepare_event_copy
        - summarize_creator_metrics
        - payment_lookup
        - provider_health_summary
        - moderation_queue_summary
        - draft_support_reply
        - prepare_refund_decision
        - prepare_ban_or_restriction
...
    AdminMutualsSafety:
      type: object
      required: [openReports, activeMutuals, staleMutuals, socialMoneyBoundary]
      properties:
        openReports: { type: integer }
        activeMutuals: { type: integer }
        staleMutuals: { type: integer }
        socialMoneyBoundary: { type: string, enum: [money_never_buys_people_visibility_matches_or_social_priority] }
...
    AdminMediaAsset:
      type: object
      required: [id, contentItemId, provider, providerAssetId, providerState, providerPlayable, hasPlaybackUrl, createdAt]
      properties:
        id: { type: string, format: uuid }
        contentItemId: { type: string, format: uuid }
        provider: { type: string, enum: [bunny, livepeer] }
        providerAssetId: { type: string }
        providerState: { type: string }
        providerPlayable: { type: boolean }
        hasPlaybackUrl: { type: boolean }
        readyAt: { type: string, format: date-time, nullable: true }
        providerCheckedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    AdminAgeCheck:
      type: object
      required: [id, userId, provider, providerReference, state, hasProviderReference, privacyBoundary, createdAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        provider: { type: string }
        providerReference: { type: string }
        state: { $ref: "#/components/schemas/AgeState" }
        jurisdiction: { type: string, nullable: true }
        rule: { type: string, nullable: true }
        hasProviderReference: { type: boolean }
        privacyBoundary: { type: string, enum: [sanitized_age_state_no_raw_identity_payloads] }
        verifiedAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    AdminIdentityCheck:
      type: object
      required: [id, userId, provider, providerReference, verificationType, state, hasProviderReference, hasLegalNameHash, privacyBoundary, createdAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        provider: { type: string }
        providerReference: { type: string }
        verificationType: { type: string, enum: [age, identity, liveness, wallet_ownership, kyc, kyb] }
        state: { type: string, enum: [pending, verified, failed, expired, cancelled] }
        countryCode: { type: string, nullable: true }
        documentType: { type: string, nullable: true }
        livenessState: { type: string, nullable: true }
        walletOwnershipState: { type: string, nullable: true }
        hasProviderReference: { type: boolean }
        hasLegalNameHash: { type: boolean }
        privacyBoundary: { type: string, enum: [sanitized_identity_minimized_no_raw_documents_or_pii] }
        verifiedAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    AdminAiSession:
      type: object
      required: [id, actorUserId, scope, state, allowedToolCount, createdAt, expiresAt]
      properties:
        id: { type: string, format: uuid }
        actorUserId: { type: string, format: uuid }
        scope: { type: string, enum: [user_self_service, creator_helper, admin_ops] }
        state: { type: string, enum: [active, expired, revoked] }
        allowedToolCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    AdminAiToolCall:
      type: object
      required: [id, sessionId, actorUserId, scope, toolName, state, confirmationState, inputSummary, outputSummary, redactionBoundary, createdAt]
      properties:
        id: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        actorUserId: { type: string, format: uuid }
        scope: { type: string, enum: [user_self_service, creator_helper, admin_ops] }
        toolName: { $ref: "#/components/schemas/AiToolName" }
        state: { type: string, enum: [prepared, executed, blocked, failed] }
        confirmationState: { type: string, enum: [not_required, required, confirmed, rejected] }
        subjectType: { type: string, enum: [content, event, payment, provider, support_case, report, user, none], nullable: true }
        subjectId: { type: string, nullable: true }
        inputSummary: { type: string }
        outputSummary: { type: string }
        redactionBoundary: { type: string, enum: [summaries_only_no_tool_payloads_or_secrets] }
        createdAt: { type: string, format: date-time }
...